backend/pyproject.toml
[project]
name = "sentilex-ai"
version = "0.1.0"
description = "Backend for the legally-compliant, forensically-secure AI system for Sri Lankan Law."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.4",
    "ipykernel>=7.1.0",
    "langchain>=1.2.7",
    "langchain-community>=0.2.0",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=1.1.7",
    "langchain-pinecone>=0.2.13",
    "mcp[cli]>=1.26.0",
    "psycopg2-binary>=2.9.9",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
    "boto3>=1.42.36",
    "langgraph>=1.0.7",
    "openai>=2.15.0",
    "tiktoken>=0.12.0",
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "pydantic-core>=2.41.5",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "email-validator>=2.3.0",
    "pyjwt>=2.10.1",
    "cryptography>=46.0.4",
    "argon2-cffi>=23.1.0",
    "pyotp>=2.9.0",
    "qrcode>=7.4.2",
    "pillow>=10.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiosmtplib>=3.0.1",
    "jinja2>=3.1.3",
    "itsdangerous>=2.1.2",
    "slowapi>=0.1.9",
    "user-agents>=2.2.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation>=0.60b1",
    "prometheus-client>=0.24.1",
    "jsonschema>=4.26.0",
    "pyyaml>=6.0.3",
    "ormsgpack>=1.12.1",
    "orjson>=3.11.5",
    "rich>=14.3.1",
    "pywin32>=311 ; sys_platform == 'win32'",
    "sendgrid==6.11.0",
    "weasyprint>=60.0",
    "zipstream-ng>=1.7.0",
]
